// src/app/core/interfaces/JovenInvestigadorDTO.ts
export interface JovenInvestigadorDTO {

  tipoIdentificacion?: string;
  identificacion?: string;
  lugarExpedicion?: string;
  apellido1?: string;
  apellido2?: string;
  nombrePila?: string;
  ciudadNacimiento?: string;
  departamentoNacimiento?: string;
  paisNacimiento?: string;
  fechaNacimiento?: string;
  municipio?: string;
  direccion?: string;
  barrio?: string;
  estrato?: number;
  tipoVivienda?: string;
  contactoPrincipal?: string;
  celular?: string;
  correoElectronico?: string;
  banco?: string;
  cuenta?: string;
  tipoCuenta?: string;
  fechaModificaParticipante?: string;
  confirmado?: string;
  confirmadoConvenio?: string;
  dependencia?: string;
  resumenActividades?: string;
  identificadorParticipante?: number;
  nombreTutor?: string;
  identificacionTutor?: string;
  correoTutor?: string;
  contactoTutor?: string;
  programa?: string;
  semestre?: number;
  fechaInicioSemestre?: string;
  fechaFinSemestre?: string;
  nombreGrupoTutor?: string;
}